+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.6.1-hf.2
+BUG FIXES
+* [\#641](https://github.com/binance-chain/node/pull/641) [Dex] Add max lock time in time lock plugin
+
+IMPROVEMENTS
+* [\#638](https://github.com/binance-chain/node/pull/638) [Pub] BEP39 - add memo to transfer kafka message
+* [\#639](https://github.com/binance-chain/node/pull/639) [ABCI] add levels parameter to depth ABCI query
+
 ## 0.6.1-hf.1
 BUG FIXES
 * [\#635](https://github.com/binance-chain/node/pull/635) fix panic in pre-check is not recovered